Module ethereum_gen::beefy_light_client

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::__shared_types::*;

Structs

AddKnownMMRRootCall

Container type for all input parameters for the addKnownMMRRoot function with signature addKnownMMRRoot(bytes32) and selector 0x58c03034

AddKnownMMRRootReturn

Container type for all return fields from the addKnownMMRRoot function with signature addKnownMMRRoot(bytes32) and selector 0x58c03034

BeefyLightClient

BeefyMMRLeaf

BeefyMMRLeaf(uint8,uint32,uint64,uint32,bytes32,bytes32,bytes32,bytes32)

CheckCommitmentSignaturesThresholdCall

Container type for all input parameters for the checkCommitmentSignaturesThreshold function with signature checkCommitmentSignaturesThreshold(uint256,uint256[]) and selector 0x0131477c

CheckValidatorInSetCall

Container type for all input parameters for the checkValidatorInSet function with signature checkValidatorInSet(address,uint256,bytes32[]) and selector 0x5d325d71

CheckValidatorInSetReturn

Container type for all return fields from the checkValidatorInSet function with signature checkValidatorInSet(address,uint256,bytes32[]) and selector 0x5d325d71

Commitment

Commitment(bytes,bytes32,bytes,uint32,uint64)

CreateCommitmentHashCall

Container type for all input parameters for the createCommitmentHash function with signature createCommitmentHash((bytes,bytes32,bytes,uint32,uint64)) and selector 0x9dbf76fb

CreateCommitmentHashReturn

Container type for all return fields from the createCommitmentHash function with signature createCommitmentHash((bytes,bytes32,bytes,uint32,uint64)) and selector 0x9dbf76fb

CreateInitialBitfieldCall

Container type for all input parameters for the createInitialBitfield function with signature createInitialBitfield(uint256[],uint256) and selector 0x5da57fe9

CreateInitialBitfieldReturn

Container type for all return fields from the createInitialBitfield function with signature createInitialBitfield(uint256[],uint256) and selector 0x5da57fe9

CreateRandomBitfieldCall

Container type for all input parameters for the createRandomBitfield function with signature createRandomBitfield(uint256[],uint256) and selector 0x74de3a6f

CreateRandomBitfieldReturn

Container type for all return fields from the createRandomBitfield function with signature createRandomBitfield(uint256[],uint256) and selector 0x74de3a6f

CurrentValidatorSetCall

Container type for all input parameters for the currentValidatorSet function with signature currentValidatorSet() and selector 0x2cdea717

CurrentValidatorSetReturn

Container type for all return fields from the currentValidatorSet function with signature currentValidatorSet() and selector 0x2cdea717

EncodeMMRLeafCall

Container type for all input parameters for the encodeMMRLeaf function with signature encodeMMRLeaf((uint8,uint32,uint64,uint32,bytes32,bytes32,bytes32,bytes32)) and selector 0x1b7a2c7a

EncodeMMRLeafReturn

Container type for all return fields from the encodeMMRLeaf function with signature encodeMMRLeaf((uint8,uint32,uint64,uint32,bytes32,bytes32,bytes32,bytes32)) and selector 0x1b7a2c7a

ErrorAndSafetyBufferCall

Container type for all input parameters for the ERROR_AND_SAFETY_BUFFER function with signature ERROR_AND_SAFETY_BUFFER() and selector 0xbe7e93a3

ErrorAndSafetyBufferReturn

Container type for all return fields from the ERROR_AND_SAFETY_BUFFER function with signature ERROR_AND_SAFETY_BUFFER() and selector 0xbe7e93a3

GetLatestMMRRootCall

Container type for all input parameters for the getLatestMMRRoot function with signature getLatestMMRRoot() and selector 0x970ca0d9

GetLatestMMRRootReturn

Container type for all return fields from the getLatestMMRRoot function with signature getLatestMMRRoot() and selector 0x970ca0d9

HashMMRLeafCall

Container type for all input parameters for the hashMMRLeaf function with signature hashMMRLeaf(bytes) and selector 0xf4fa4e45

HashMMRLeafReturn

Container type for all return fields from the hashMMRLeaf function with signature hashMMRLeaf(bytes) and selector 0xf4fa4e45

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(uint64,(uint128,uint128,bytes32),(uint128,uint128,bytes32)) and selector 0xe104815d

IsKnownRootCall

Container type for all input parameters for the isKnownRoot function with signature isKnownRoot(bytes32) and selector 0x6d9833e3

IsKnownRootReturn

Container type for all return fields from the isKnownRoot function with signature isKnownRoot(bytes32) and selector 0x6d9833e3

LatestBeefyBlockCall

Container type for all input parameters for the latestBeefyBlock function with signature latestBeefyBlock() and selector 0x66ae69a0

LatestBeefyBlockReturn

Container type for all return fields from the latestBeefyBlock function with signature latestBeefyBlock() and selector 0x66ae69a0

LatestMMRRootIndexCall

Container type for all input parameters for the latestMMRRootIndex function with signature latestMMRRootIndex() and selector 0x74d21421

LatestMMRRootIndexReturn

Container type for all return fields from the latestMMRRootIndex function with signature latestMMRRootIndex() and selector 0x74d21421

LatestMMRRootsCall

Container type for all input parameters for the latestMMRRoots function with signature latestMMRRoots(uint256) and selector 0x5aebc629

LatestMMRRootsReturn

Container type for all return fields from the latestMMRRoots function with signature latestMMRRoots(uint256) and selector 0x5aebc629

LatestRandomSeedCall

Container type for all input parameters for the latestRandomSeed function with signature latestRandomSeed() and selector 0xbc5f084a

LatestRandomSeedReturn

Container type for all return fields from the latestRandomSeed function with signature latestRandomSeed() and selector 0xbc5f084a

MaximumBlockGapCall

Container type for all input parameters for the MAXIMUM_BLOCK_GAP function with signature MAXIMUM_BLOCK_GAP() and selector 0x4afad95b

MaximumBlockGapReturn

Container type for all return fields from the MAXIMUM_BLOCK_GAP function with signature MAXIMUM_BLOCK_GAP() and selector 0x4afad95b

MmrRootHistorySizeCall

Container type for all input parameters for the MMR_ROOT_HISTORY_SIZE function with signature MMR_ROOT_HISTORY_SIZE() and selector 0x7954d0d0

MmrRootHistorySizeReturn

Container type for all return fields from the MMR_ROOT_HISTORY_SIZE function with signature MMR_ROOT_HISTORY_SIZE() and selector 0x7954d0d0

MmrRootIdCall

Container type for all input parameters for the MMR_ROOT_ID function with signature MMR_ROOT_ID() and selector 0x0a7c8faa

MmrRootIdReturn

Container type for all return fields from the MMR_ROOT_ID function with signature MMR_ROOT_ID() and selector 0x0a7c8faa

MmrVerificationCall

Container type for all input parameters for the mmrVerification function with signature mmrVerification() and selector 0x801ed1e3

MmrVerificationReturn

Container type for all return fields from the mmrVerification function with signature mmrVerification() and selector 0x801ed1e3

NewMMRRootFilter

NextValidatorSetCall

Container type for all input parameters for the nextValidatorSet function with signature nextValidatorSet() and selector 0x36667513

NextValidatorSetReturn

Container type for all return fields from the nextValidatorSet function with signature nextValidatorSet() and selector 0x36667513

NumberOfBlocksPerSessionCall

Container type for all input parameters for the NUMBER_OF_BLOCKS_PER_SESSION function with signature NUMBER_OF_BLOCKS_PER_SESSION() and selector 0xe2a6ff3f

NumberOfBlocksPerSessionReturn

Container type for all return fields from the NUMBER_OF_BLOCKS_PER_SESSION function with signature NUMBER_OF_BLOCKS_PER_SESSION() and selector 0xe2a6ff3f

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

RequiredNumberOfSignaturesCall

Container type for all input parameters for the requiredNumberOfSignatures function with signature requiredNumberOfSignatures() and selector 0x72fe1a9f

RequiredNumberOfSignaturesReturn

Container type for all return fields from the requiredNumberOfSignatures function with signature requiredNumberOfSignatures() and selector 0x72fe1a9f

RequiredNumberOfSignaturesWithNumValidatorsCall

Container type for all input parameters for the requiredNumberOfSignatures function with signature requiredNumberOfSignatures(uint256) and selector 0x6edda8f4

RequiredNumberOfSignaturesWithNumValidatorsReturn

Container type for all return fields from the requiredNumberOfSignatures function with signature requiredNumberOfSignatures(uint256) and selector 0x6edda8f4

ResetCall

Container type for all input parameters for the reset function with signature reset(uint64,(uint128,uint128,bytes32),(uint128,uint128,bytes32)) and selector 0x8d5d1fae

SubmitSignatureCommitmentCall

Container type for all input parameters for the submitSignatureCommitment function with signature submitSignatureCommitment((bytes,bytes32,bytes,uint32,uint64),(uint256[],bytes[],uint256[],address[],bytes32[][]),(uint8,uint32,uint64,uint32,bytes32,bytes32,bytes32,bytes32),(bytes32[],uint64)) and selector 0x73d02f8e

ThresholdDenominatorCall

Container type for all input parameters for the THRESHOLD_DENOMINATOR function with signature THRESHOLD_DENOMINATOR() and selector 0xef024458

ThresholdDenominatorReturn

Container type for all return fields from the THRESHOLD_DENOMINATOR function with signature THRESHOLD_DENOMINATOR() and selector 0xef024458

ThresholdNumeratorCall

Container type for all input parameters for the THRESHOLD_NUMERATOR function with signature THRESHOLD_NUMERATOR() and selector 0x5a8d2f0e

ThresholdNumeratorReturn

Container type for all return fields from the THRESHOLD_NUMERATOR function with signature THRESHOLD_NUMERATOR() and selector 0x5a8d2f0e

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

ValidatorProof

ValidatorProof(uint256[],bytes[],uint256[],address[],bytes32[][])

ValidatorSet

ValidatorSet(uint128,uint128,bytes32)

VerificationSuccessfulFilter

VerifyBeefyMerkleLeafCall

Container type for all input parameters for the verifyBeefyMerkleLeaf function with signature verifyBeefyMerkleLeaf(bytes32,(bytes32[],uint64)) and selector 0x2d268c9e

VerifyBeefyMerkleLeafReturn

Container type for all return fields from the verifyBeefyMerkleLeaf function with signature verifyBeefyMerkleLeaf(bytes32,(bytes32[],uint64)) and selector 0x2d268c9e

VerifyNewestMMRLeafCall

Container type for all input parameters for the verifyNewestMMRLeaf function with signature verifyNewestMMRLeaf((uint8,uint32,uint64,uint32,bytes32,bytes32,bytes32,bytes32),bytes32,(bytes32[],uint64)) and selector 0x3226d9f4

Enums

BeefyLightClientCalls

Container type for all of the contract’s call

BeefyLightClientEvents

Container type for all of the contract’s events

Statics

BEEFYLIGHTCLIENT_ABI

The parsed JSON ABI of the contract.