Module ethereum_gen::beefylightclient_mod

Re-exports

pub use super::__shared_types::*;

Structs

BeefyLightClient

BeefyMMRLeaf

BeefyMMRLeaf(uint8,uint32,bytes32,uint64,uint32,bytes32,bytes32)

BlockWaitPeriodCall

Container type for all input parameters for the BLOCK_WAIT_PERIOD function with signature BLOCK_WAIT_PERIOD() and selector [251, 117, 44, 98]

Commitment

Commitment(bytes,bytes32,bytes,uint32,uint64)

CompleteSignatureCommitmentCall

Container type for all input parameters for the completeSignatureCommitment function with signature completeSignatureCommitment(uint256,(bytes,bytes32,bytes,uint32,uint64),(bytes[],uint256[],address[],bytes32[][]),(uint8,uint32,bytes32,uint64,uint32,bytes32,bytes32),(bytes32[],uint64)) and selector [45, 201, 37, 224]

CreateCommitmentHashCall

Container type for all input parameters for the createCommitmentHash function with signature createCommitmentHash((bytes,bytes32,bytes,uint32,uint64)) and selector [157, 191, 118, 251]

CreateInitialBitfieldCall

Container type for all input parameters for the createInitialBitfield function with signature createInitialBitfield(uint256[],uint256) and selector [93, 165, 127, 233]

CreateRandomBitfieldCall

Container type for all input parameters for the createRandomBitfield function with signature createRandomBitfield(uint256) and selector [146, 132, 128, 22]

CurrentIdCall

Container type for all input parameters for the currentId function with signature currentId() and selector [224, 13, 209, 97]

EncodeMMRLeafCall

Container type for all input parameters for the encodeMMRLeaf function with signature encodeMMRLeaf((uint8,uint32,bytes32,uint64,uint32,bytes32,bytes32)) and selector [55, 185, 135, 228]

ErrorAndSafetyBufferCall

Container type for all input parameters for the ERROR_AND_SAFETY_BUFFER function with signature ERROR_AND_SAFETY_BUFFER() and selector [190, 126, 147, 163]

FinalVerificationSuccessfulFilter

HashMMRLeafCall

Container type for all input parameters for the hashMMRLeaf function with signature hashMMRLeaf(bytes) and selector [244, 250, 78, 69]

InitialVerificationSuccessfulFilter

LatestBeefyBlockCall

Container type for all input parameters for the latestBeefyBlock function with signature latestBeefyBlock() and selector [102, 174, 105, 160]

LatestMMRRootCall

Container type for all input parameters for the latestMMRRoot function with signature latestMMRRoot() and selector [65, 201, 99, 78]

MaximumBlockGapCall

Container type for all input parameters for the MAXIMUM_BLOCK_GAP function with signature MAXIMUM_BLOCK_GAP() and selector [74, 250, 217, 91]

MmrRootIdCall

Container type for all input parameters for the MMR_ROOT_ID function with signature MMR_ROOT_ID() and selector [10, 124, 143, 170]

MmrVerificationCall

Container type for all input parameters for the mmrVerification function with signature mmrVerification() and selector [128, 30, 209, 227]

NewMMRRootFilter

NewSignatureCommitmentCall

Container type for all input parameters for the newSignatureCommitment function with signature newSignatureCommitment(bytes32,uint256[],bytes,uint256,address,bytes32[]) and selector [229, 77, 21, 67]

NumberOfBlocksPerSessionCall

Container type for all input parameters for the NUMBER_OF_BLOCKS_PER_SESSION function with signature NUMBER_OF_BLOCKS_PER_SESSION() and selector [226, 166, 255, 63]

RequiredNumberOfSignaturesCall

Container type for all input parameters for the requiredNumberOfSignatures function with signature requiredNumberOfSignatures() and selector [114, 254, 26, 159]

RequiredNumberOfSignaturesWithNumValidatorsCall

Container type for all input parameters for the requiredNumberOfSignatures function with signature requiredNumberOfSignatures(uint256) and selector [110, 221, 168, 244]

ThresholdDenominatorCall

Container type for all input parameters for the THRESHOLD_DENOMINATOR function with signature THRESHOLD_DENOMINATOR() and selector [239, 2, 68, 88]

ThresholdNumeratorCall

Container type for all input parameters for the THRESHOLD_NUMERATOR function with signature THRESHOLD_NUMERATOR() and selector [90, 141, 47, 14]

ValidationDataCall

Container type for all input parameters for the validationData function with signature validationData(uint256) and selector [32, 191, 165, 203]

ValidatorProof

ValidatorProof(bytes[],uint256[],address[],bytes32[][])

ValidatorRegistryCall

Container type for all input parameters for the validatorRegistry function with signature validatorRegistry() and selector [243, 118, 235, 187]

VerifyBeefyMerkleLeafCall

Container type for all input parameters for the verifyBeefyMerkleLeaf function with signature verifyBeefyMerkleLeaf(bytes32,(bytes32[],uint64)) and selector [45, 38, 140, 158]

VerifyNewestMMRLeafCall

Container type for all input parameters for the verifyNewestMMRLeaf function with signature verifyNewestMMRLeaf((uint8,uint32,bytes32,uint64,uint32,bytes32,bytes32),bytes32,(bytes32[],uint64)) and selector [36, 61, 38, 95]

Enums

BeefyLightClientCalls

BeefyLightClientEvents

Statics

BEEFYLIGHTCLIENT_ABI