Module ethereum_gen::bridge

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AcceptedEthTokensCall

Container type for all input parameters for the acceptedEthTokens function with signature acceptedEthTokens(address) and selector 0x687a82db

AcceptedEthTokensReturn

Container type for all return fields from the acceptedEthTokens function with signature acceptedEthTokens(address) and selector 0x687a82db

AddEthNativeTokenCall

Container type for all input parameters for the addEthNativeToken function with signature addEthNativeToken(address,string,string,uint8,bytes32,uint8[],bytes32[],bytes32[]) and selector 0xdfd9db89

AddNewSidechainTokenCall

Container type for all input parameters for the addNewSidechainToken function with signature addNewSidechainToken(string,string,uint8,bytes32,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x23ad7de2

AddPeerByPeerCall

Container type for all input parameters for the addPeerByPeer function with signature addPeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0xca70cf6e

AddPeerByPeerReturn

Container type for all return fields from the addPeerByPeer function with signature addPeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0xca70cf6e

AddressVALCall

Container type for all input parameters for the _addressVAL function with signature _addressVAL() and selector 0x3b698288

AddressVALReturn

Container type for all return fields from the _addressVAL function with signature _addressVAL() and selector 0x3b698288

AddressXORCall

Container type for all input parameters for the _addressXOR function with signature _addressXOR() and selector 0xc29bb332

AddressXORReturn

Container type for all return fields from the _addressXOR function with signature _addressXOR() and selector 0xc29bb332

Bridge

ChangePeersFilter

DepositFilter

IsPeerCall

Container type for all input parameters for the isPeer function with signature isPeer(address) and selector 0x3e44cf78

IsPeerReturn

Container type for all return fields from the isPeer function with signature isPeer(address) and selector 0x3e44cf78

MigratedFilter

NetworkIdCall

Container type for all input parameters for the _networkId function with signature _networkId() and selector 0x5fde87c5

NetworkIdReturn

Container type for all return fields from the _networkId function with signature _networkId() and selector 0x5fde87c5

PeersCountCall

Container type for all input parameters for the peersCount function with signature peersCount() and selector 0xae6664e0

PeersCountReturn

Container type for all return fields from the peersCount function with signature peersCount() and selector 0xae6664e0

PrepareForMigrationCall

Container type for all input parameters for the prepareForMigration function with signature prepareForMigration(bytes32,uint8[],bytes32[],bytes32[]) and selector 0x75ff82bd

ReceiveByEthereumAssetAddressCall

Container type for all input parameters for the receiveByEthereumAssetAddress function with signature receiveByEthereumAssetAddress(address,uint256,address,address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x135f14e2

ReceiveBySidechainAssetIdCall

Container type for all input parameters for the receiveBySidechainAssetId function with signature receiveBySidechainAssetId(bytes32,uint256,address,address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x75273ece

ReceivePaymentCall

Container type for all input parameters for the receivePayment function with signature receivePayment() and selector 0x835c19f3

RemovePeerByPeerCall

Container type for all input parameters for the removePeerByPeer function with signature removePeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x89c39baf

RemovePeerByPeerReturn

Container type for all return fields from the removePeerByPeer function with signature removePeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x89c39baf

SendERC20ToSidechainCall

Container type for all input parameters for the sendERC20ToSidechain function with signature sendERC20ToSidechain(bytes32,uint256,address) and selector 0x03ef9d1f

SendEthToSidechainCall

Container type for all input parameters for the sendEthToSidechain function with signature sendEthToSidechain(bytes32) and selector 0xfdfba25d

ShutDownAndMigrateCall

Container type for all input parameters for the shutDownAndMigrate function with signature shutDownAndMigrate(bytes32,address,address[],uint8[],bytes32[],bytes32[]) and selector 0xdab4a4c6

SidechainTokenAddressArrayCall

Container type for all input parameters for the _sidechainTokenAddressArray function with signature _sidechainTokenAddressArray(uint256) and selector 0x16283134

SidechainTokenAddressArrayReturn

Container type for all return fields from the _sidechainTokenAddressArray function with signature _sidechainTokenAddressArray(uint256) and selector 0x16283134

SidechainTokensByAddressCall

Container type for all input parameters for the _sidechainTokensByAddress function with signature _sidechainTokensByAddress(address) and selector 0xcd550947

SidechainTokensByAddressReturn

Container type for all return fields from the _sidechainTokensByAddress function with signature _sidechainTokensByAddress(address) and selector 0xcd550947

SidechainTokensCall

Container type for all input parameters for the _sidechainTokens function with signature _sidechainTokens(bytes32) and selector 0x8505e868

SidechainTokensReturn

Container type for all return fields from the _sidechainTokens function with signature _sidechainTokens(bytes32) and selector 0x8505e868

UniqueAddressesCall

Container type for all input parameters for the _uniqueAddresses function with signature _uniqueAddresses(address) and selector 0x12c6fa1b

UniqueAddressesReturn

Container type for all return fields from the _uniqueAddresses function with signature _uniqueAddresses(address) and selector 0x12c6fa1b

UsedCall

Container type for all input parameters for the used function with signature used(bytes32) and selector 0xb07c411f

UsedReturn

Container type for all return fields from the used function with signature used(bytes32) and selector 0xb07c411f

WithdrawalFilter

Enums

BridgeCalls

Container type for all of the contract’s call

BridgeEvents

Container type for all of the contract’s events

Statics

BRIDGE_ABI

The parsed JSON ABI of the contract.