Module ethereum_gen::inbound_channel

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Re-exports

pub use super::__shared_types::*;

Structs

Batch

Batch(uint256,uint256,(address,uint256,bytes)[])

BatchDispatchedFilter

BatchNonceCall

Container type for all input parameters for the batch_nonce function with signature batch_nonce() and selector 0xbc27722b

BatchNonceReturn

Container type for all return fields from the batch_nonce function with signature batch_nonce() and selector 0xbc27722b

BeefyLightClientCall

Container type for all input parameters for the beefyLightClient function with signature beefyLightClient() and selector 0xaf41c33e

BeefyLightClientReturn

Container type for all return fields from the beefyLightClient function with signature beefyLightClient() and selector 0xaf41c33e

ConfigUpdateRoleCall

Container type for all input parameters for the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector 0xa2d6c6e5

ConfigUpdateRoleReturn

Container type for all return fields from the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector 0xa2d6c6e5

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

DefaultAdminRoleReturn

Container type for all return fields from the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

GasBufferCall

Container type for all input parameters for the GAS_BUFFER function with signature GAS_BUFFER() and selector 0xc7e67360

GasBufferReturn

Container type for all return fields from the GAS_BUFFER function with signature GAS_BUFFER() and selector 0xc7e67360

GasProofsCall

Container type for all input parameters for the gas_proofs function with signature gas_proofs(uint64) and selector 0x4ce1e4db

GasProofsReturn

Container type for all return fields from the gas_proofs function with signature gas_proofs(uint64) and selector 0x4ce1e4db

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GetRoleAdminReturn

Container type for all return fields from the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector 0x2f2ff15d

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

HasRoleReturn

Container type for all return fields from the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

InboundChannel

InitializeCall

Container type for all input parameters for the initialize function with signature initialize() and selector 0x8129fc1c

LeafBytes

LeafBytes(bytes,bytes,bytes)

Message

Message(address,uint256,bytes)

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector 0x36568abe

ResetCall

Container type for all input parameters for the reset function with signature reset() and selector 0xd826f88f

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector 0xd547741f

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SubmitCall

Container type for all input parameters for the submit function with signature submit((uint256,uint256,(address,uint256,bytes)[]),(bytes,bytes,bytes),(bytes32[],uint64)) and selector 0xe54af507

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

SupportsInterfaceReturn

Container type for all return fields from the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

Enums

InboundChannelCalls

Container type for all of the contract’s call

InboundChannelEvents

Container type for all of the contract’s events

Statics

INBOUNDCHANNEL_ABI

The parsed JSON ABI of the contract.