Module ethereum_gen::incentivizedinboundchannel_mod

Re-exports

pub use super::__shared_types::*;

Structs

Batch

Batch(uint256,(address,uint64,uint256,uint256,bytes)[])

BeefyLightClientCall

Container type for all input parameters for the beefyLightClient function with signature beefyLightClient() and selector [175, 65, 195, 62]

ConfigUpdateRoleCall

Container type for all input parameters for the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector [162, 214, 198, 229]

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector [162, 23, 253, 223]

GasBufferCall

Container type for all input parameters for the GAS_BUFFER function with signature GAS_BUFFER() and selector [199, 230, 115, 96]

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector [36, 138, 156, 163]

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector [47, 47, 241, 93]

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector [145, 209, 72, 84]

IncentivizedInboundChannel

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address,address) and selector [72, 92, 201, 85]

LeafBytes

LeafBytes(bytes,bytes,bytes)

Message

Message(address,uint64,uint256,uint256,bytes)

MessageDispatchedFilter

NonceCall

Container type for all input parameters for the nonce function with signature nonce() and selector [175, 254, 208, 224]

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector [54, 86, 138, 190]

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector [213, 71, 116, 31]

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SubmitCall

Container type for all input parameters for the submit function with signature submit((uint256,(address,uint64,uint256,uint256,bytes)[]),(bytes,bytes,bytes),(bytes32[],uint64)) and selector [45, 187, 160, 214]

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector [1, 255, 201, 167]

Enums

IncentivizedInboundChannelCalls

IncentivizedInboundChannelEvents

Statics

INCENTIVIZEDINBOUNDCHANNEL_ABI