Module ethereum_gen::incentivizedoutboundchannel_mod

Structs

AuthorizeDefaultOperatorCall

Container type for all input parameters for the authorizeDefaultOperator function with signature authorizeDefaultOperator(address) and selector [183, 66, 164, 4]

AuthorizeOperatorCall

Container type for all input parameters for the authorizeOperator function with signature authorizeOperator(address) and selector [149, 155, 140, 63]

ConfigUpdateRoleCall

Container type for all input parameters for the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector [162, 214, 198, 229]

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector [162, 23, 253, 223]

FeeCall

Container type for all input parameters for the fee function with signature fee() and selector [221, 202, 63, 67]

FeeChangedFilter

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector [36, 138, 156, 163]

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector [47, 47, 241, 93]

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector [145, 209, 72, 84]

IncentivizedOutboundChannel

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address[],address[],uint256) and selector [246, 108, 0, 39]

IsOperatorForCall

Container type for all input parameters for the isOperatorFor function with signature isOperatorFor(address,address) and selector [217, 91, 99, 113]

MessageFilter

NonceCall

Container type for all input parameters for the nonce function with signature nonce() and selector [175, 254, 208, 224]

OperatorAuthorizedFilter

OperatorRevokedFilter

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector [54, 86, 138, 190]

RevokeDefaultOperatorCall

Container type for all input parameters for the revokeDefaultOperator function with signature revokeDefaultOperator(address) and selector [42, 117, 52, 163]

RevokeOperatorCall

Container type for all input parameters for the revokeOperator function with signature revokeOperator(address) and selector [250, 216, 179, 42]

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector [213, 71, 116, 31]

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SetFeeCall

Container type for all input parameters for the setFee function with signature setFee(uint256) and selector [105, 254, 14, 45]

SubmitCall

Container type for all input parameters for the submit function with signature submit(address,bytes) and selector [118, 132, 110, 221]

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector [1, 255, 201, 167]

Enums

IncentivizedOutboundChannelCalls

IncentivizedOutboundChannelEvents

Statics

INCENTIVIZEDOUTBOUNDCHANNEL_ABI