Module ethereum_gen::master

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddPeerByPeerCall

Container type for all input parameters for the addPeerByPeer function with signature addPeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0xca70cf6e

AddPeerByPeerReturn

Container type for all return fields from the addPeerByPeer function with signature addPeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0xca70cf6e

AddTokenCall

Container type for all input parameters for the addToken function with signature addToken(address) and selector 0xd48bfca7

CheckTokenAddressCall

Container type for all input parameters for the checkTokenAddress function with signature checkTokenAddress(address) and selector 0x9f1a156c

CheckTokenAddressReturn

Container type for all return fields from the checkTokenAddress function with signature checkTokenAddress(address) and selector 0x9f1a156c

EnableContractFilter

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address,address[],string,string,uint8,address,uint256,uint256,bytes32) and selector 0x5c6d0b8c

IsOwnerCall

Container type for all input parameters for the isOwner function with signature isOwner() and selector 0x8f32d59b

IsOwnerReturn

Container type for all return fields from the isOwner function with signature isOwner() and selector 0x8f32d59b

IsPeerCall

Container type for all input parameters for the isPeer function with signature isPeer(address) and selector 0x3e44cf78

IsPeerReturn

Container type for all return fields from the isPeer function with signature isPeer(address) and selector 0x3e44cf78

IsTokenCall

Container type for all input parameters for the isToken function with signature isToken(address) and selector 0x19f37361

IsTokenReturn

Container type for all return fields from the isToken function with signature isToken(address) and selector 0x19f37361

Master

MintTokensByPeersCall

Container type for all input parameters for the mintTokensByPeers function with signature mintTokensByPeers(address,uint256,address,bytes32,uint8[],bytes32[],bytes32[],address) and selector 0x78749620

OwnerCall

Container type for all input parameters for the owner_ function with signature owner_() and selector 0xe7663079

OwnerReturn

Container type for all return fields from the owner_ function with signature owner_() and selector 0xe7663079

PeersCountCall

Container type for all input parameters for the peersCount function with signature peersCount() and selector 0xae6664e0

PeersCountReturn

Container type for all return fields from the peersCount function with signature peersCount() and selector 0xae6664e0

ProofCall

Container type for all input parameters for the proof function with signature proof() and selector 0xfaf924cf

ProofReturn

Container type for all return fields from the proof function with signature proof() and selector 0xfaf924cf

ProofRewardCall

Container type for all input parameters for the proofReward function with signature proofReward() and selector 0x92d75299

ProofRewardReturn

Container type for all return fields from the proofReward function with signature proofReward() and selector 0x92d75299

RecieveCall

Container type for all input parameters for the recieve function with signature recieve() and selector 0xa9e10bf2

RemovePeerByPeerCall

Container type for all input parameters for the removePeerByPeer function with signature removePeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x89c39baf

RemovePeerByPeerReturn

Container type for all return fields from the removePeerByPeer function with signature removePeerByPeer(address,bytes32,uint8[],bytes32[],bytes32[]) and selector 0x89c39baf

SubmitProofCall

Container type for all input parameters for the submitProof function with signature submitProof(bytes32,uint8[],bytes32[],bytes32[]) and selector 0x74e808af

TokenInstanceCall

Container type for all input parameters for the tokenInstance function with signature tokenInstance() and selector 0x658030b3

TokenInstanceReturn

Container type for all return fields from the tokenInstance function with signature tokenInstance() and selector 0x658030b3

UniqueAddressesCall

Container type for all input parameters for the uniqueAddresses function with signature uniqueAddresses(address) and selector 0x1d345ebb

UniqueAddressesReturn

Container type for all return fields from the uniqueAddresses function with signature uniqueAddresses(address) and selector 0x1d345ebb

UsedCall

Container type for all input parameters for the used function with signature used(bytes32) and selector 0xb07c411f

UsedReturn

Container type for all return fields from the used function with signature used(bytes32) and selector 0xb07c411f

WithdrawCall

Container type for all input parameters for the withdraw function with signature withdraw(address,uint256,address,bytes32,uint8[],bytes32[],bytes32[],address) and selector 0xeea29e3e

WithdrawalFilter

Enums

MasterCalls

Container type for all of the contract’s call

MasterEvents

Container type for all of the contract’s events

Statics

MASTER_ABI

The parsed JSON ABI of the contract.