Module ethereum_gen::migration_app

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

DefaultAdminRoleReturn

Container type for all return fields from the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GetRoleAdminReturn

Container type for all return fields from the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector 0x2f2ff15d

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

HasRoleReturn

Container type for all return fields from the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

InboundCall

Container type for all input parameters for the inbound function with signature inbound() and selector 0x4a026227

InboundChannelRoleCall

Container type for all input parameters for the INBOUND_CHANNEL_ROLE function with signature INBOUND_CHANNEL_ROLE() and selector 0xbc0c22a7

InboundChannelRoleReturn

Container type for all return fields from the INBOUND_CHANNEL_ROLE function with signature INBOUND_CHANNEL_ROLE() and selector 0xbc0c22a7

InboundReturn

Container type for all return fields from the inbound function with signature inbound() and selector 0x4a026227

MigrateEthCall

Container type for all input parameters for the migrateEth function with signature migrateEth(address) and selector 0x175dc507

MigrateNativeErc20Call

Container type for all input parameters for the migrateNativeErc20 function with signature migrateNativeErc20(address,address[]) and selector 0xd0c902a7

MigrateSidechainCall

Container type for all input parameters for the migrateSidechain function with signature migrateSidechain(address,address[]) and selector 0x8032147f

MigratedEthFilter

MigratedNativeErc20Filter

MigratedSidechainFilter

MigrationApp

OutboundCall

Container type for all input parameters for the outbound function with signature outbound() and selector 0x5109932e

OutboundReturn

Container type for all return fields from the outbound function with signature outbound() and selector 0x5109932e

ReceivePaymentCall

Container type for all input parameters for the receivePayment function with signature receivePayment() and selector 0x835c19f3

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector 0x36568abe

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector 0xd547741f

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

SupportsInterfaceReturn

Container type for all return fields from the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

Enums

MigrationAppCalls

Container type for all of the contract’s call

MigrationAppEvents

Container type for all of the contract’s events

Statics

MIGRATIONAPP_ABI

The parsed JSON ABI of the contract.