Module ethereum_gen::migrationapp_mod

Structs

ChannelsCall

Container type for all input parameters for the channels function with signature channels(uint8) and selector [204, 211, 65, 50]

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector [162, 23, 253, 223]

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector [36, 138, 156, 163]

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector [47, 47, 241, 93]

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector [145, 209, 72, 84]

InboundChannelRoleCall

Container type for all input parameters for the INBOUND_CHANNEL_ROLE function with signature INBOUND_CHANNEL_ROLE() and selector [188, 12, 34, 167]

MigrateEthCall

Container type for all input parameters for the migrateEth function with signature migrateEth(address) and selector [23, 93, 197, 7]

MigrateNativeErc20Call

Container type for all input parameters for the migrateNativeErc20 function with signature migrateNativeErc20(address,address[]) and selector [208, 201, 2, 167]

MigrateSidechainCall

Container type for all input parameters for the migrateSidechain function with signature migrateSidechain(address,address[]) and selector [128, 50, 20, 127]

MigratedEthFilter

MigratedNativeErc20Filter

MigratedSidechainFilter

MigrationApp

ReceivePaymentCall

Container type for all input parameters for the receivePayment function with signature receivePayment() and selector [131, 92, 25, 243]

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector [54, 86, 138, 190]

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector [213, 71, 116, 31]

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector [1, 255, 201, 167]

Enums

MigrationAppCalls

MigrationAppEvents

Statics

MIGRATIONAPP_ABI