Module ethereum_gen::outbound_channel

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AuthorizeDefaultOperatorCall

Container type for all input parameters for the authorizeDefaultOperator function with signature authorizeDefaultOperator(address) and selector 0xb742a404

AuthorizeOperatorCall

Container type for all input parameters for the authorizeOperator function with signature authorizeOperator(address) and selector 0x959b8c3f

ConfigUpdateRoleCall

Container type for all input parameters for the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector 0xa2d6c6e5

ConfigUpdateRoleReturn

Container type for all return fields from the CONFIG_UPDATE_ROLE function with signature CONFIG_UPDATE_ROLE() and selector 0xa2d6c6e5

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

DefaultAdminRoleReturn

Container type for all return fields from the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

FeeCall

Container type for all input parameters for the fee function with signature fee() and selector 0xddca3f43

FeeChangedFilter

FeeReturn

Container type for all return fields from the fee function with signature fee() and selector 0xddca3f43

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GetRoleAdminReturn

Container type for all return fields from the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector 0x2f2ff15d

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

HasRoleReturn

Container type for all return fields from the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address[],address[],uint256) and selector 0xf66c0027

IsOperatorForCall

Container type for all input parameters for the isOperatorFor function with signature isOperatorFor(address,address) and selector 0xd95b6371

IsOperatorForReturn

Container type for all return fields from the isOperatorFor function with signature isOperatorFor(address,address) and selector 0xd95b6371

MessageFilter

NonceCall

Container type for all input parameters for the nonce function with signature nonce() and selector 0xaffed0e0

NonceReturn

Container type for all return fields from the nonce function with signature nonce() and selector 0xaffed0e0

OperatorAuthorizedFilter

OperatorRevokedFilter

OutboundChannel

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector 0x36568abe

ResetCall

Container type for all input parameters for the reset function with signature reset() and selector 0xd826f88f

RevokeDefaultOperatorCall

Container type for all input parameters for the revokeDefaultOperator function with signature revokeDefaultOperator(address) and selector 0x2a7534a3

RevokeOperatorCall

Container type for all input parameters for the revokeOperator function with signature revokeOperator(address) and selector 0xfad8b32a

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector 0xd547741f

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SetFeeCall

Container type for all input parameters for the setFee function with signature setFee(uint256) and selector 0x69fe0e2d

SubmitCall

Container type for all input parameters for the submit function with signature submit(address,bytes) and selector 0x76846edd

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

SupportsInterfaceReturn

Container type for all return fields from the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

Enums

OutboundChannelCalls

Container type for all of the contract’s call

OutboundChannelEvents

Container type for all of the contract’s events

Statics

OUTBOUNDCHANNEL_ABI

The parsed JSON ABI of the contract.