Module ethereum_gen::sidechainapp_mod

Structs

BurnedFilter

ChannelsCall

Container type for all input parameters for the channels function with signature channels(uint8) and selector [204, 211, 65, 50]

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector [162, 23, 253, 223]

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector [36, 138, 156, 163]

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector [47, 47, 241, 93]

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector [145, 209, 72, 84]

InboundChannelRoleCall

Container type for all input parameters for the INBOUND_CHANNEL_ROLE function with signature INBOUND_CHANNEL_ROLE() and selector [188, 12, 34, 167]

LockCall

Container type for all input parameters for the lock function with signature lock(address,bytes32,uint256,uint8) and selector [73, 227, 127, 19]

MintedFilter

RegisterAssetCall

Container type for all input parameters for the registerAsset function with signature registerAsset(string,string,bytes32) and selector [136, 2, 165, 65]

RegisterExistingAssetCall

Container type for all input parameters for the registerExistingAsset function with signature registerExistingAsset(address) and selector [72, 224, 240, 87]

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector [54, 86, 138, 190]

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector [213, 71, 116, 31]

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SidechainApp

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector [1, 255, 201, 167]

TokensCall

Container type for all input parameters for the tokens function with signature tokens(address) and selector [228, 134, 3, 57]

UnlockCall

Container type for all input parameters for the unlock function with signature unlock(address,bytes32,address,uint256) and selector [1, 12, 227, 199]

Enums

SidechainAppCalls

SidechainAppEvents

Statics

SIDECHAINAPP_ABI