Module ethereum_gen::test_token

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AllowanceCall

Container type for all input parameters for the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

AllowanceReturn

Container type for all return fields from the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

ApprovalFilter

ApproveCall

Container type for all input parameters for the approve function with signature approve(address,uint256) and selector 0x095ea7b3

ApproveReturn

Container type for all return fields from the approve function with signature approve(address,uint256) and selector 0x095ea7b3

BalanceOfCall

Container type for all input parameters for the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BalanceOfReturn

Container type for all return fields from the balanceOf function with signature balanceOf(address) and selector 0x70a08231

DecimalsCall

Container type for all input parameters for the decimals function with signature decimals() and selector 0x313ce567

DecimalsReturn

Container type for all return fields from the decimals function with signature decimals() and selector 0x313ce567

DecreaseAllowanceCall

Container type for all input parameters for the decreaseAllowance function with signature decreaseAllowance(address,uint256) and selector 0xa457c2d7

DecreaseAllowanceReturn

Container type for all return fields from the decreaseAllowance function with signature decreaseAllowance(address,uint256) and selector 0xa457c2d7

IncreaseAllowanceCall

Container type for all input parameters for the increaseAllowance function with signature increaseAllowance(address,uint256) and selector 0x39509351

IncreaseAllowanceReturn

Container type for all return fields from the increaseAllowance function with signature increaseAllowance(address,uint256) and selector 0x39509351

MintCall

Container type for all input parameters for the mint function with signature mint(address,uint256) and selector 0x40c10f19

NameCall

Container type for all input parameters for the name function with signature name() and selector 0x06fdde03

NameReturn

Container type for all return fields from the name function with signature name() and selector 0x06fdde03

SymbolCall

Container type for all input parameters for the symbol function with signature symbol() and selector 0x95d89b41

SymbolReturn

Container type for all return fields from the symbol function with signature symbol() and selector 0x95d89b41

TestToken

TotalSupplyCall

Container type for all input parameters for the totalSupply function with signature totalSupply() and selector 0x18160ddd

TotalSupplyReturn

Container type for all return fields from the totalSupply function with signature totalSupply() and selector 0x18160ddd

TransferCall

Container type for all input parameters for the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

TransferFilter

TransferFromCall

Container type for all input parameters for the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferFromReturn

Container type for all return fields from the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferReturn

Container type for all return fields from the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

Enums

TestTokenCalls

Container type for all of the contract’s call

TestTokenEvents

Container type for all of the contract’s events

Statics

TESTTOKEN_ABI

The parsed JSON ABI of the contract.