Module ethereum_gen::testtoken_mod

Structs

AllowanceCall

Container type for all input parameters for the allowance function with signature allowance(address,address) and selector [221, 98, 237, 62]

ApprovalFilter

ApproveCall

Container type for all input parameters for the approve function with signature approve(address,uint256) and selector [9, 94, 167, 179]

BalanceOfCall

Container type for all input parameters for the balanceOf function with signature balanceOf(address) and selector [112, 160, 130, 49]

DecimalsCall

Container type for all input parameters for the decimals function with signature decimals() and selector [49, 60, 229, 103]

DecreaseAllowanceCall

Container type for all input parameters for the decreaseAllowance function with signature decreaseAllowance(address,uint256) and selector [164, 87, 194, 215]

IncreaseAllowanceCall

Container type for all input parameters for the increaseAllowance function with signature increaseAllowance(address,uint256) and selector [57, 80, 147, 81]

MintCall

Container type for all input parameters for the mint function with signature mint(address,uint256) and selector [64, 193, 15, 25]

NameCall

Container type for all input parameters for the name function with signature name() and selector [6, 253, 222, 3]

SymbolCall

Container type for all input parameters for the symbol function with signature symbol() and selector [149, 216, 155, 65]

TestToken

TotalSupplyCall

Container type for all input parameters for the totalSupply function with signature totalSupply() and selector [24, 22, 13, 221]

TransferCall

Container type for all input parameters for the transfer function with signature transfer(address,uint256) and selector [169, 5, 156, 187]

TransferFilter

TransferFromCall

Container type for all input parameters for the transferFrom function with signature transferFrom(address,address,uint256) and selector [35, 184, 114, 221]

Enums

TestTokenCalls

TestTokenEvents

Statics

TESTTOKEN_ABI